+++ psutil/_psutil_common.h
@@ -134,4 +134,6 @@ int psutil_setup(void);
     PVOID psutil_GetProcAddress(LPCSTR libname, LPCSTR procname);
     PVOID psutil_GetProcAddressFromLib(LPCSTR libname, LPCSTR procname);
     PVOID psutil_SetFromNTStatusErr(NTSTATUS Status, const char *syscall);
+    double psutil_FiletimeToUnixTime(FILETIME ft);
+    double psutil_LargeIntegerToUnixTime(LARGE_INTEGER li);
 #endif